requirements.txt
requests
python-dotenv
oaaclient